pyproject.toml
[tool.poetry]
name = "housemonitor"
version = "0.1.0"
description = "measure and store temperature and humidity."
authors = ["Thomas Golding"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-firestore = "^1.9.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
ipython = "^7.19.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
